+++ R/utils.R
@@ -1,0 +1,4 @@
+
+`%||%` <- function(a, b) { # nolint
+  if (is.null(a)) b else a
+}