+++ NEWS.md
@@ -1,3 +1,8 @@
+# gasworks 0.1.0
+
+* add stochastic compartmental model of GAS transmission and pmcmc fitting
+capabilities via mrc-ide/mcstate.git
+
 # gasworks 0.0.1
 
 * initial setup